pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-companion"
version = "0.1.0"
description = "An interactive AI companion with a visual avatar and voice capabilities."
authors = [{name = "AI Companion Team", email = "contact@ai-companion.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["AI", "companion", "avatar", "voice", "chatbot", "llm", "tts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
    "speechrecognition>=3.10.0",
    "pyaudio>=0.2.11",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "huggingface-hub>=0.16.0",
    "llama-cpp-python>=0.2.0",
    "kokoro-tts>=0.1.0; platform_machine != 'aarch64'",
    "webrtcvad>=2.0.10",
    "librosa>=0.10.0",
    "scipy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-companion"
Repository = "https://github.com/yourusername/ai-companion"
Issues = "https://github.com/yourusername/ai-companion/issues"

[project.scripts]
ai-companion = "ai_companion.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
